+++ NEWS.md
@@ -1,10 +1,13 @@
 # FitzRoy 0.1.0.9000
 
-* Added function
+* Added historical data for footywire.com
+* Added historical data for afltable.com
+* Added function to get advanced player stats
+* Added function to get fixture data
+* Added function to get results data
 
 # FitzRoy 0.0.0.9000
 
 * Created package
 * Added a `NEWS.md` file to track changes to the package.
-* Added historical data for footywire.com
-* Added historical data for afltable.com
+