+++ NEWS.md
@@ -1,10 +1,12 @@
 # FitzRoy 0.1.0.9000
+Initial release of FitzRoy package. 
 
 * Added historical data for footywire.com
 * Added historical data for afltable.com
 * Added function to get advanced player stats
 * Added function to get fixture data
 * Added function to get results data
+* Added 2017 weather data
 
 # FitzRoy 0.0.0.9000